Fix forceCopy to check backup exists before delete

+++ system/ee/ExpressionEngine/Library/Filesystem/Filesystem.php
@@ -423,7 +423,9 @@ class Filesystem
             throw $e;
         }
 
-        $this->delete($destBackup);
+        if ($this->exists($destBackup)) {
+            $this->delete($destBackup);
+        }
 
     }
